Article management.

+++ haxorbb/profile/views.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from . import profile
 from .. import db
-from flask import (current_app, url_for, redirect, render_template, Response, send_file)
+from flask import (current_app, url_for, redirect, render_template, flash, send_file)
 from werkzeug import secure_filename
 from ..models import User
 from .forms import Profile, Upload, Rename
@@ -10,6 +10,7 @@ from datetime import datetime, timedelta
 import os
 from PIL import Image
 from ..utilities.filters import create_timg
+import pytz
 
 try:
     from os import scandir
@@ -37,14 +38,20 @@ def view(username):
         return redirect('front_page.home_page')
 
 
+def build_timezone_list():
+    return [(tz, tz) for tz in pytz.common_timezones]
+
+
 @profile.route('/view/<username>/edit', methods=['GET', 'POST'])
 @login_required
 def edit_profile(username):
-    if current_user.username != username and not current_user.is_administrator:
+    if current_user.username != username or not current_user.is_administrator:
         return redirect(url_for('front_page.home_page'))
     user = User.query.filter_by(username=username).first()
     file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', user.username)
     form = Profile()
+    form.time_zone.choices = build_timezone_list()
+
     # Check if the user is using 2FA, and needs to auth.
     if user.otp:
         tfa_state = True
@@ -55,12 +62,15 @@ def edit_profile(username):
         user.location = form.location.data
         user.avatar_text = form.avatar_text.data
         user.avatar_url = form.avatar_url.data
+        user.timezone = form.time_zone.data
         db.session.add(user)
         db.session.commit()
+
     form.fullname.data = user.fullname or None
     form.location.data = user.location or None
     form.avatar_url.data = user.avatar_url or None
     form.avatar_text.data = user.avatar_text or None
+    form.time_zone.data = user.timezone
     file_list = [f.stat().st_size for f in scandir(file_path)]
     disk_use = sum(file_list)
     return render_template('profile/edit.html', user=user, form=form, tfa=tfa_state, disk_use=disk_use)
@@ -69,10 +79,9 @@ def edit_profile(username):
 @profile.route('/view/<username>/files', methods=['GET', 'POST'])
 @login_required
 def manage_files(username):
-    filedata = []
-    if current_user.username != username and not current_user.is_administrator:
+    if current_user.username != username or not current_user.is_administrator:
         return redirect(url_for('front_page.home_page'))
-
+    filedata = []
     user = User.query.filter_by(username=username).first()
     file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', user.username)
     if not os.path.isdir(file_path):
@@ -103,8 +112,9 @@ def manage_files(username):
 @profile.route('/view/<username>/files/upload', methods=['GET', 'POST'])
 @login_required
 def user_upload(username):
-    if current_user.username != username and not current_user.is_administrator:
+    if current_user.username != username or not current_user.is_administrator:
         return redirect(url_for('front_page.home_page'))
+    user = User.query.filter_by(username=username).first()
     file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', username)
     ALLOWED_EXTENSIONS = ['png', 'jpg', 'gif', 'jpeg']
     form = Upload()
@@ -113,13 +123,25 @@ def user_upload(username):
         filename = secure_filename(file_data.filename)
         if filename.rsplit('.')[1] in ALLOWED_EXTENSIONS:
             file_data.save(os.path.join(file_path, filename))
+        else:
+            flash("Unaccecptable file type submitted for upload")
+            return redirect(url_for('profile.manage_files', username=username))
+        flash("Filed uploaded successfully")
         return redirect(url_for('profile.manage_files', username=username))
-    return render_template('profile/upload.html', user=username, form=form)
+
+    return render_template('profile/upload.html', username=username, form=form, user=user)
+
+
+@profile.errorhandler(413)
+def request_entity_too_large(error):
+    return "File exceeds upload limits", 413
 
 
 @profile.route('/view/<username>/files/rename/<filename>', methods=['GET', 'POST'])
 @login_required
 def rename_file(username, filename):
+    if current_user.username != username or not current_user.is_administrator:
+        return redirect(url_for('front_page.home_page'))
     form = Rename()
     if form.validate_on_submit():
         file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', username)
@@ -132,7 +154,9 @@ def rename_file(username, filename):
 
 @profile.route('/view/<username>/files/delete/<filename>', methods=['GET'])
 @login_required
-def delete_file(username ,filename):
+def delete_file(username, filename):
+    if current_user.username != username or not current_user.is_administrator:
+        return redirect(url_for('front_page.home_page'))
     file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', username, filename)
     os.remove(file_path)
     return redirect(url_for('profile.manage_files', username=username))
@@ -141,6 +165,8 @@ def delete_file(username ,filename):
 @profile.route('/view/<username>/files/<filename>/set_avatar', methods=['GET'])
 @login_required
 def set_avatar(username, filename):
+    if current_user.username != username or not current_user.is_administrator:
+        return redirect(url_for('front_page.home_page'))
     user = User.query.filter_by(username=username).first()
     new_avatar = url_for('media', filename='users/{}/{}'.format(user.username, filename))
     user.set_avatar_url(new_avatar)
@@ -150,6 +176,8 @@ def set_avatar(username, filename):
 @profile.route('/view/<username>/files/<filename>/set_picture', methods=['GET'])
 @login_required
 def set_picture(username, filename):
+    if current_user.username != username or not current_user.is_administrator:
+        return redirect(url_for('front_page.home_page'))
     user = User.query.filter_by(username=username).first()
     new_picture = url_for('media', filename='users/{}/{}'.format(user.username, filename))
     user.set_picture(new_picture)
@@ -159,6 +187,8 @@ def set_picture(username, filename):
 @profile.route('/view/<username>/download/<filename>')
 @login_required
 def download_file(username, filename):
+    if current_user.username != username or not current_user.is_administrator:
+        return redirect(url_for('front_page.home_page'))
     file_path = os.path.join(current_app.config['MEDIA_ROOT'], 'users', username, filename)
     img_type = filename.rsplit('.')[1]
     return send_file(file_path, mimetype='image/{}'.format(img_type), as_attachment=True)
